requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.9.1
dj-database-url==2.0.0
Django==5.2
django-cors-headers==4.3.0
django-filter==25.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.5.1
drf-yasg==1.21.7
factory-boy==3.2.1
Faker==37.4.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
psycopg2-binary==2.9.7
PyJWT==2.7.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
